collector_parquet/requirements.txt
# ============================================
# FILE: requirements.txt (UPDATED)
# ============================================

# Core dependencies
python-dotenv
websockets

# Data processing
pandas
pyarrow